requirements.txt
pdf2image
chromadb
pandas
requests
pillow
pytesseract
langchain
langchain-community
faiss-cpu
python-dotenv
fastai
fastapi 
uvicorn [standard]
python-multipart

opencv-python
numpy
regex
PyYAML
loguru
reportlab
qrcode
pydantic
striprtf
python-dotenv